✅ test: split isSafeEnumerationOption test to sibling file
The new predicate test pushed GBNFGrammarBuilderTests.swift over swiftlint's
400-line file_length cap. Move it to a sibling extension file per the
testing.md convention (extension, not a new @Suite).
